+++ patches/net/minecraft/world/item/Item.java.patch
@@ -35,7 +35,7 @@
 +    /** @deprecated Neo: do not use, use {@link net.neoforged.neoforge.event.ModifyDefaultComponentsEvent the event} instead */
 +    @org.jetbrains.annotations.ApiStatus.Internal @Deprecated
 +    public void modifyDefaultComponentsFrom(net.minecraft.core.component.DataComponentPatch patch) {
-+        if (!net.neoforged.neoforge.registries.RegistryManager.canModifyComponents()) throw new IllegalStateException("Default components cannot be modified now!");
++        if (!net.neoforged.neoforge.internal.RegistrationEvents.canModifyComponents()) throw new IllegalStateException("Default components cannot be modified now!");
 +        var builder = DataComponentMap.builder().addAll(components);
 +        patch.entrySet().forEach(entry -> builder.set((DataComponentType)entry.getKey(), entry.getValue().orElse(null)));
 +        var built = builder.build();
