+++ src/main/java/net/neoforged/neoforge/registries/RegistryManager.java
@@ -25,7 +25,6 @@ import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.ResourceLocation;
 import net.neoforged.fml.ModLoader;
-import net.neoforged.neoforge.event.ModifyDefaultComponentsEvent;
 import net.neoforged.neoforge.network.configuration.RegistryDataMapNegotiation;
 import net.neoforged.neoforge.network.handling.IPayloadContext;
 import net.neoforged.neoforge.network.payload.FrozenRegistryPayload;
@@ -95,18 +94,6 @@ public class RegistryManager {
                     + pendingModdedRegistries.stream().map(ResourceLocation::toString).collect(Collectors.joining("\n\t - ", "\n\t - ", "")));
         }
         pendingModdedRegistries = null;
-    }
-
-    private static boolean canModifyComponents;
-
-    public static void modifyComponents() {
-        canModifyComponents = true;
-        ModLoader.postEvent(new ModifyDefaultComponentsEvent());
-        canModifyComponents = false;
-    }
-
-    public static boolean canModifyComponents() {
-        return canModifyComponents;
     }
 
     public static void initDataMaps() {
